+++ packages/vm/src/lib.rs
@@ -5,7 +5,6 @@ mod backend;
 mod cache;
 mod calls;
 mod capabilities;
-mod checksum;
 mod compatibility;
 mod conversion;
 mod environment;
@@ -41,7 +40,6 @@ pub use crate::calls::{
     call_ibc_packet_receive_raw, call_ibc_packet_timeout, call_ibc_packet_timeout_raw,
 };
 pub use crate::capabilities::capabilities_from_csv;
-pub use crate::checksum::Checksum;
 pub use crate::errors::{
     CommunicationError, CommunicationResult, RegionValidationError, RegionValidationResult,
     VmError, VmResult,